Refactor CreateClient
CreateClient involves creating a clientset that accesses the k8 api
It does not belong in chaos but rather kubernetes
