+++ kubemonkey/kubemonkey.go
@@ -10,6 +10,7 @@ import (
 	"github.com/asobti/kube-monkey/config"
 	"github.com/asobti/kube-monkey/calendar"
 	"github.com/asobti/kube-monkey/schedule"
+	"github.com/asobti/kube-monkey/kubernetes"
 )
 
 func durationToNextRun(runhour int, location *time.Location) time.Duration {
@@ -28,7 +29,7 @@ func durationToNextRun(runhour int, location *time.Location) time.Duration {
 func Run() error {
 	// Verify kubernetes client can be created and works before
 	// we enter execution loop
-	if _, err := chaos.CreateClient(); err != nil {
+	if _, err := kubernetes.CreateClient(); err != nil {
 		return err
 	}
